Disallow prefixes on type annotations
I had a prefixed type at the top-level let previously, but it leads to
a bad case that breaks idempotency in the formatter (see parent commit).
It is possible to remediate this while preserving prefixes, but that
would add complexity, and why would do you even want to have comments
between the : and the type? I allowed the prefixes in type lists, in
generic instantiations and in function types, because there I you might
want to document the types, e.g.

    let f: (
      // The number of widgets.
      Int,
      // The maximum widget serial number.
      Int,
    ) -> Int = (n, m) => ...

But between the : and the type ... then you should just put the comment
above the let. So we simplify the CST, and fix a bug!