Add initial static typechecker
This is a somewhat big and invasive change, so let's pause here and
commit. The typechecker is incomplete, it doesn't handle many AST nodes,
and it is also not used yet.

At some point while doing this, I realized that the type for Type, with
the Rc's for the inner types, is not right. There was &Rc<Type> in some
places which was awkward, and in some places I then had to box (rc) a
primitive type which was silly, I want to pass just &Type. So I tried
making all nested Type members of Type unboxed, but then boxing the
contents of the type itself for complex types. Mid-way through the
change it was a mess everywhere and I wasn't sure this was better, but I
continued and the end result is way nicer and should also be more
performant. In hindsight, I should apply the same transformation to the
Value type. That should help a lot with performance too, no more double
boxing. But I can do all of that later, it will be an invasive change.