Define a type lattice after all
I didn't want to go down the path of unifying all the elements of alist
type, because I fear it will be expensive when loading large json
documents with many big and deep dicts. Inferring those will lead to
huge types, and we can't even share the Rc instance if they are inferred
all the time.

BUT, let's not worry about performance right now. Setting that concern
aside, it is very tempting to just do the inference. And then a lattice
naturally falls out. Maybe in the end I end up removing some of this
again, but just having the lattice will probably lead to a better
design.

Then an open question: I have meet, which is enough for primitive types
and for ??variant types, (I never remember, is it covariant or
contravariant?), which List and Dict and Set are in RCL, because you can
only get stuff out, not put stuff in. Function return types also behave
like that. But then for function arguments I need join. And join of two
incompatible things is Void. Which makes sense in a statically typed
setting. If I do e.g.

    let fs = [
      // (Int) -> Int
      x => x + 1,
      // (Bool) -> Bool
      x => not x
    ]

Then in general an element of 'fs' is a function that I cannot call
because the input would need to be both Int and Bool. But at runtime, I
don't call a functino "in general", I call a particular one, which can
be fine. How to type that? I think that if any of the arguments
collapsed to Void, we need to instead make it Dynamic, and check at
runtime. But that is the complete opposite of what the lattice does! It
feels inelegant and like a hack. So I'm not sure about that yet, but
let's see when we get to it, maybe I'm missing something obvious and
I'll have a better idea then.