Implement simple type expr evaluator
I already had one in eval, and it still exists. This one is simpler.
The one in the evaluator was prematurely made to use environments, to
support definitions and user-defined types. But that should be an
improvement to add *later*, right now I need to do the simplest thing I
can get away with.

The type expr evaluator in crate::eval still exists, but it should be
removed at some point.