Implement typecheck for sequences
This copies and adapts the machinery from the evaluator that tried to
infer the collection type at runtime. Much nicer to do it statically,
although the possibility of type annotations does complicate things.

The runtime evaluator checks could be removed now, but I'll do that in a
later commit. We could also add specialized AST nodes for the known
collection type, and make the typechecker write them. And finally we can
also produce empty sets, though that needs one more runtime tweak. Right
now it can infer a set type but produce a dict at runtime.