Add idea to output depfile
I want to generate some files using RCL, so how do you automate all of
that ... using Make, but then you have to specify the dependencies to
make it rebuild.

But now that I think of it, maybe it should not be Make, RCL is fast
enough (famous last words) that we can just regenerate the entire thing
all the time maybe.