Update golden output for new type error messages
Not all of them are improvements, sometimes the runtime error can give
better context than the static type error. I need to edit the type
expectation machinery to allow a _reason_ for the expectation, but we
can do that at some later time. For now let's get all the tests to pass.