Add goldens to seed fuzz corpus for types
These goldens add no new coverage, and I would like to add a comment to
explain their purpose ... but that would defeat their purpose. I let the
fuzzer run for more than an hour, and despite discovering some code
paths in the typechecker, it didn't discover all of them. So help it a
bit by creating small seeds that cover the expected cases.

One additional problem may have been that even an almost minimal example
for these is 40 bytes in the dict case, and I was fuzzing with a max
length of 32. Perhaps fuzzing with a larger input size would also help
to discover the code paths sooner.