Use the shared Env type for values
This is a backport of a larger refactor that I'm making in the
typechecker. This change turned out to be huge, so I'm breaking it down
into pieces. One of those pieces is having an Env type that can be used
both for types and for values.

The change that I'm backporting this from, first started using the new
Env type for types and then later also for values. But in this case, we
apply it to values first, and I'll rebase the typechecker, so it will
start to use this type later. But it will be compatible, and unified.