Add some more ideas
I was thinking today about how the | operator is way briefer than
writing a comprehension in a collection, but writing the comprehension
is nice to format and | is awkward. Also when writing down the type of
the | operator ... it's a challenge (also to typecheck it).

Perhaps, adding a spread/unpack operator will resolve both issues. It
would be slightly more verbose, but not terribly so. And I believe it
would be more readable too.