Make the parser not eat some comments
There were some cases where non-code before a closing delimiter could be
eaten, because we consumed it before discovering the closing delimiter.
This happens in the real world, mostly when you have a collection and
you comment out the last element. To fix this, put the non-code in the
CST and emit it when formatting.