Begin workin on Python module
Pyo3 looks okay, it brings in quite some dependencies but it's not as
outrageous as e.g. Tokio. Having to use this Maturin thing ... I haven't
gotten it to work yet, is it really needed? For packaging as a Python
wheel it's nice, but if I just want to try it locally?

Anyway, let's commit what we have so far.