Add a Rust eDSL to annotate stdlib with types
I need to typecheck built-in methods and functions, so I need to know
their types. Constructing the type ASTs in Rust by hand would get really
really tedious, so add some macros to make them a bit more pleasant to
write and read. The macros get a bit ugly though, but macros tend to do
that, it's not too bad if you start reading through them.

Unfortunately the types cannot be consts, because they contain Vecs and
Rcs, so make them a function that returns the types.