Set a more aggressive evaluation limit when fuzzing
The fuzzer discovered a timeout, which was a nested loop over two
large-ish std.range calls. Yeah, if you materialize a Cartesian product,
that is going to be expensive, but that's a "doctor it hurts when I
press here -- then don't press there" problem and we already have a
mitigation for it. Tune the limit of that mitigation a bit for fuzzing,
so it doesn't waste as much time on inputs that time out like the nested
loop.

Update: I also bumped the limit for the non-fuzzing use case so it can
evaluate one of my Advent of Code solutions without hitting the limit.
It's a stretch use case anyway, but having one that works is nice.