Add an empty Tree-sitter grammar
The thing that will unlock the widest range of syntax highlighting, is
probably a Tree Sitter grammar. Unfortunately this has to be written in
Javascript and involves node.js, but fortunately, only for the initial
setup and generate step, so hopefully I can get away with not needing
the node.js dependency in most places and just commit the resulting
grammar to the repository. Maybe there can be one check in the flake
that ensures it's up to date.

Maybe this Tree Sitter grammar can even replace the Bison one if it's
readable. But let's see how it goes.

At first I also included the generated parser. But after writing all of
this, I rebased it out. I don't want it to pollute my diffs. People who
want to use the grammar will have to run the generator. Maybe I could
commit the grammar.json ... But for now, no generated files in the repo.