Handle Void type in the typechecker
Things keep getting uglier and uglier, with more and more special cases.
But if you have an empty collection and you try to index into it, yeah,
that is guaranteed to fail, and the typechecker already knows that, no
need to execute the program. Still, it would probably be clearer for the
user to just execute the program.