Prevent another case of Dynamic in type error
This one was discovered by the fuzzer. The proper solution is to wrap
the node in a typecheck, but we don't have the node there, so at this
point we still rely on dynamic checks.

At this point the typechecker is becoming so ad-hoc. I need to find
a better way to do this, and it probably involves Hindley-Milner and
constraint propagation.