Document how to compile tree-sitter grammar
I was struggling with this before, but now I found the problem:
tree-sitter calls CC, and I had CC=clang, but somehow then I get an
executable that links against some files in the Nix store, but also
against a libstdc++.so.6 which is nowhere to be found. When I set CC
to a GCC that Nix provides, everything works fine.