Unbox primitive values
In short: this flips runtime::Value from being used as Rc<Value> in most
places to just Value, with the Rc pushed inside for compound and large
values.

This is a thing I realized when working on the type system. For Type, I
initially had a similar representation as for Value, but then I
discovered the unboxed form which is simpler, more elegant, and more
performant.

It is a better representation for Value too. Don't box them everywhere,
but only box the internals for big values. Then e.g. a List of Int will
just have the ints inline, instead of all on the heap. Also we don't
double-box strings.

This refactor is worth it just for the simplifications, but I expected
it to have a performance benefit as well. If it does, it is not very
obvious. In a very rough benchmark (no attempt at eliminating noise at
all), there is a tiny difference, but it doesn't look significant to me.
The benchmark is to load a large-ish json document (a ListenBrainz
export). Perhaps the impact is less because the time is dominated by
string formatting. Both binaries are compiled in release mode.

Benchmark 1: /tmp/rcl-boxed eval listenbrainz.json
  Time (mean ± σ):     15.470 s ±  0.525 s    [User: 13.970 s, System: 1.471 s]
  Range (min … max):   14.966 s … 16.495 s    10 runs

Benchmark 2: /tmp/rcl-unboxed eval listenbrainz.json
  Time (mean ± σ):     15.180 s ±  0.264 s    [User: 13.758 s, System: 1.398 s]
  Range (min … max):   14.804 s … 15.486 s    10 runs

Summary
  /tmp/rcl-unboxed eval listenbrainz.json ran
    1.02 ± 0.04 times faster than /tmp/rcl-boxed eval listenbrainz.json