Define an Env type for evaluator and typecheck
Both the evaluator and the typechecker need to track the variables that
are in scope, and a way to do lookups, etc. Currently this is handled in
the runtime Env type, and in my first iteration of the typechecker I
extended it to hold both types and values.

A more elegant way is to define an Env type for tracking any associated
value (a runtime value for the evaluator, a type for the typechecker),
that can be used by both.

I finished this in my static typechecker branch. But that change is
getting huge. So I'm trying to break it up into smaller changes, and one
of the logical changes that might be feasible to extract already, is to
use this shared Env type. So if it comes out of the blue in this commit
and it's not used yet ... it was developed in a different context.