Port some of function types to the new check
This is incomplete. For cases where there is a type requirement, I still
have to check the arity and put the right guarantees in the environment.
And also, I feel the TypeReq is missing from the runtime function type.
If there is anything to verify before the call, then we need to know
where the requirements came from! But I guess I'll run into that as soon
as I get to typechecking calls and that will get sorted out.