Correct goldens for desirable new error messages
With the new typechecker, a few error messages are now different and it
changes the goldens again. These messages are good as they are now,
accept the new version. There are also some regarding function arity and
such that have regressed, I still need to implement more of the static
checks there.