Define SourcedType, rewrite the typechecker again
After rewriting the entire typechecker with TypeReq, at the point where
I got to functions and contravariance, everything started breaking down
again. So here is attempt III, again with a single type type, and a
check for subtypes. But after writing down some things on paper, I
realized that the subtype check should not return a boolean. That's not
the meaning of subtype that I want. So when you make subtyping a partial
ordering (surprise surprise, it is just inclusion on sets, which is also
a partial order but not a total order), the case where two types are
unordered is exactly the case where I need to insert a runtime type
check!

This time I did rewrite the entire thing in one go without trying to
keep it compiling in the meantime, so this is a rather big commit.
Apologies. Everything compiles now, but the goldens need to be updated
once more.