Bring back Void type error, track sources
I ran into exactly the same case I fixed before, trying to print Dynamic
in an error because Void was expected. Good I added the regression test!
Now we have a nicer way of dealing with it, too. Although I'm not so
sure the error is an improvement, it all starts to be extremely rigid,
and while it's nice to detect errors like indexing into an empty dict
that is guaranteed to fail, the error can be more intractable than just
letting it fail.