Rename "Dynamic" to "Any"
After going through the lattice exercise, and clarifying to myself what
that means, "dynamic" is now no longer a special type that gets treated
differently from the others (like it is, or at least looked to me, in C#
where I took the name from). It is just the type that describes all
possible values, and "Any" is a very good name for that. Also, the long
"Dynamic" was starting to annoy me because it's too long, and also, this
makes it consistent with Mypy.