Document that the typechecker may reject programs
I was thinking about this, with my subtype relation. If I really want
"not (T < U)" to mean "every possible value of T is not an instance of
U", then "List[Int]" and "List[String]" might actually be compatible,
and I need a runtime check. But that misses the point of the type
system!

Also add one of these as a golden. I want the other one too, but I first
need to implement reporting nested type errors.