Unify arity checks and error reporting
Runtime and typecheck-time arity checks can now use the same machinery
from the function signature. This started out as me trying to remove the
duplicated argument names, to ensure there are as many names as types by
construction, and it ended with replacing all of the arity checks that
was one of the remaining big changes of the static typechecker.

One additional change here is that we now track the spans of where the
arguments are defined, so we can include them in the error message. Not
just the function definition, but the proper span. Also, this now works
both at typecheck time and at runtime!

One regression here: built-in functions no longer perform the arity
check, which means they could go out of bounds when trying to access an
argument. I'll fix that up in the next commit. The tree is currently
unsafe!