Simplify and fix call frames in calls from builtins
I went through two iterations of this commit, original message, which is
invalid, below:
                             *   *   *

I don't know what I was thinking when I wrote the convoluted eval_call
in the evaluator. We don't need closures and whatnot. Just map_err is
sufficient. The calls happen in Rust's native stack order, so we will
append the call frames to the error in the right order.

This should fix the error messages, that were improved with the previous
arity error reporting work, but then dropped some context (I called it
previously, but now it's clear those are just call frames). This is now
also fixed, but I'll update the goldens separately.

                             *   *   *

New conclusion: it is not an additional call frame, because eval_call
itself already adds the call frame to the error. So the original
approach of customizing the error message was correct. But we can
simplify it slightly.

On top of this, whether or not we have a custom call frame matters for
whether we want to include it or not. An arity mismatch is quite obvious
if you can see the call (no custom error, it has a span in the code),
versus from an internal call that you cannot see. So with these changes,
I think the errors *finally* make sense.