Cover more cases in identifier golden test
Also update some test code to avoid unreachable cases there. This brings
coverage of strings.rs to 100%. It's not so much of a goal to get it to
100% for the sake of the number, it's just that it's much easier to spot
regressions and discover new cases when everything was at 100% and it's
not any more, then to notice that it went from 98.7 to 98.3 and you
added one case distinction that is now untested.