Think about whether all values should be comparable
My feeling is, let's be stricter by default, not allow it, and we could
always relax it if it turns out to be useful? But saying that values are
not orderable is also kind of a lie, because you can sort heterogeneous
lists and such. But doing a compare is one step further, maybe it should
not be allowed?