Include span of actual type in error when far away
We have information about why a particular type was inferred, and that
information is useful to include in some errors. But in most cases, the
type error will already be pointing at the span where the actual type
was inferred, so it's not so helpful to include that span again, it's
just useless noise.