Add "share config between tools" as a key point
I think I only really realized today, in such a clear phrasing, that
this is possible, and how powerful it is. At work I'm often updating
our configuration of which users exist, and similar configuration lives
in five different places, but cannot be reduced to a single point of
truth because all these tools use different formats.

Of course you could write a Python program to generate all of it. But
somehow with RCL the barrier to generating stuff feels much lower.