Document Cargo as a way to install
I hate it when projects list some arcane package manager as the official
way to install. In 2015 when "cargo install" was added, I was arguing
against it for this reason. But it's now 2024, Rust adoption has grown
a lot -- at least in my bubble. We can at least mention it as an
alternative.

See also: github.com / rust-lang / rfcs/pull/1200#issuecomment-121014961
(Hyperlink defused to avoid GitHub cross-referencing it.)