Simplify call frame handling in evaluation
This time I think I got this one right. There is even the possibility to
omit the call frame for calls to BuiltinMethod and BuiltinFunction, and
then when they get called from a builtin it still works. I initially
thought that that would be useful because those call stacks are so
verbose and add little value, but now I'm not so sure any more that they
add little value, for some errors it *is* helpful to still know the name
of the thing being called, even if you already blame the argument span
and the function name is right next to it.