Be more cautious when replacing call frames
I thought of this case after writing it yesterday. Especiall an arity
error can occur, and by default we don't add a call frame, so then
replacing may be dangerous if there was a different frame on the stack
already. Though now that I'm writing this, I don't think that case can
happen.