Exclude some regions from code coverage
Saying that a Debug impl is not covered is counterproductive. It is not
expected to be covered when code is correct. So exclude these regions,
because a drop from 100% coverage to <100% stands out a lot more than a
drop from 97.8% to 97.2% when I do add one line that is not covered but
might be an important edge case.