Simplify evaluation of Seq with type info
Now that the typechecker can determine whether something is a dict or
set (or list) literal, we can use that information in the evaluator.
The complexity moved to the typechecker already, so now we reap the
benefits and simplify the evaluator.

In fact there is another opportunity to simplify the evaluator further
by eliminating the SeqOut type entirely, but I'll do that in a follow-up
commit. The SeqOut served its purpose though, because the typechecker is
based on it.