Simplify eval of Seq now that we have types
Now that the evaluator knows what kind of collection to expect, because
of the typechecker, we no longer need to carry this runtime SeqOut thing
around and match on it all the time. We can work with closures instead!
This has more advantages:

 * We don't build up vecs of keys and values when we build a dict, we
   can build the dict directly.
 * Same for sets.
 * Maybe LLVM will even inline stuff?
 * If I want to do the unique dicts that deny duplicate keys, that is
   now much easier too. Just return an error from on_assoc.

But performance aside (which I really don't know if it matters, I
haven't measured), this mainly simplifies the logic, even though the
length of the code is about the same.