Reserve capacity for list evaluation
We will likely have at least as many elements as syntactic elements.
(This is not always the case, you could have "if", but that would be
silly, I don't expect it to be common.) You can have many more with
loops, but for large (or even small) lists of static size, this could
help to avoid a few reallocations and not waste so much excess. (But
then the growth could create more excess ...)