Remove resolved question
No, I don't want to admit "_" as a type name. When we still had
"Dynamic", then yes that was too verbose. But "Any" is fine, we don't
need "_".

There should be one -- and preferably only one -- obvious way to do it.
That way was not obvious to me at first though even though I'm Dutch.