Fuzz that toml and json output can be parsed
As it turns out, inline tables in toml are not as widely allowed as I
thought. Probably it is me who misunderstood the spec because the crate
seems at least somewhat tested and this would be an easy thing to catch.
But regardless what is "right", what matters is interoperability. So
let's test the json and toml output against at least one parser.

Would be nice to also test it against Python's built-in parser, but that
will be a lot harder to fuzz.

The new/adjusted fuzzer immediately finds a counterexample that fails to
parse.