TS: Document how to run Tree-sitter tests
I wanted to add a Nix flake check to automatically run them, but
unfortunately Tree-sitter writes in arbitrary places in the file
system and I haven't found a way yet to tell it where to put the
files. So for now, only documentation it is.

This is very unfortunate because the grammar will go out of date,
I would like to test that it can parse e.g. all examples. But let's
worry about it later, maybe a differential fuzzer will be enough.