TS: Be more lenient when parsing binop chains
After some thought, this is much easier, and it's okay for the
Tree-sitter parser to be more lenient. It's main job is to highlight and
to provide useful parses even when the document is incomplete, and it
can do a better job of that when the binops chains are allowed at least
in the grammar. That you need to disambiguate is more of a -Werror style
check than a problem in the grammar.