Early out when pretty-printing a wide doc
If the result is too wide, we will switch to tall anyway, so we don't
have to finish printing the entire thing. This should make a big
difference for printing documents.

For my big json document that I test with, it doesn't make a *big*
difference, but it does cause a measurable speedup:

Benchmark 1: /tmp/before eval --color=ansi big.json
  Time (mean ± σ):     10.656 s ±  0.246 s    [User: 9.624 s, System: 1.018 s]
  Range (min … max):   10.431 s … 11.091 s    10 runs

Benchmark 2: /tmp/after eval --color=ansi big.json
  Time (mean ± σ):     10.169 s ±  0.235 s    [User: 8.846 s, System: 1.307 s]
  Range (min … max):    9.839 s … 10.429 s    10 runs

Summary
  /tmp/after eval --color=ansi big.json ran
    1.05 ± 0.03 times faster than /tmp/before eval --color=ansi big.json

This is compared to master, before Doc printing had the intermediate
stage. But as I already measured, that does not have a meaningful
impact.