Use markup in 'rcl format' output
This way 'rcl format' and 'rcl highlight' become kind of the same thing,
except that 'rcl format' has more information (it has the CST) so it can
do a better job of highlighting, just like Tree-sitter can.

(But it does require a document that is syntactically valid.)

Also it's just nicer in your terminal when the output is colored.