Struggle with box drawing glyphs in html
It renders ugly no matter what I try. The best thing would be to not
treat it as opaque monospace, and render html divs for those errors.
Then I can style the border using css. But then I need to edit the error
printing quite a lot, I don't want to do that right now. For now this is
good enough. One thing at a time.