TS: Extend string literal support
I started out with a quick hack to parse something, now let's make it
more proper. Maybe I don't need a custom scanner to keep track of the
stack depth, I think it can just be encoded in the rules! That would be
very nice, no additional external C code needed, simplifies the build
process I suppose.

This adds support for """-quoted strings and escape sequences, but there
are some issues with the """ strings for now. But let's see how much
further this gets us.

The """-strings were discovered by the fuzzer as the next low-hanging
fruit to fix.