TS: Tweak highlight queries, add test for highlight
This makes the string escapes work in both the Tree-sitter CLI and in
Helix. Also, because the parser tests are not granular enough that I
can test where the escapes are, add a highlighter test.

Edit: Or don't add a highlighter test ... it works for one assertion
line, but when I add a second one, it complains about finding a comment
scope instead of the expected scope. I fiddled with this for 15 minutes
and can't get it to work and I don't understand what I am doing wrong.
No highlight tests it is then.