TS: Drop blank node
The fuzzer discovered an input of the form "x => y" but with a blank
between the "x" and the "=>". Tree-sitter rejected that, I suppose
because blank is a token so whitespace cannot be that.

We could add blank in all places where it is allowed, but let's see how
far we get by just dropping it.