TS: Make comment an extras token
I was wondering for a long time how this works -- people are not
inserting the comment at every valid location, are they? And it turns
out no, that is what "extras" is for. I missed it in the docs.

This is also the thing that allows control over whitespace, so I can
also make the fuzzer inputs less strict now.