TS: Clear up matching immediate tokens again
With some studying the Tree-sitter Javascript parser for inspiration on
how it can be done, in the end after trying everything elese ("no that
precedence can't possibly be it"), ... it was the precedence after all.
In hindsight it makes sense, we want string tokens to bind more tightly
than comment or whitespace, and then it works out.