Add a derivation for the Tree-sitter generated files
I don't want to keep all of the generated files in the repository. But I
do want to be able to test that the fuzzers compile, so the Rust crate
needs to compile. But then I need the tree-sitter sources. So I need to
run the generate step. This is one step towards that.