Add wasm functions to load a value
I am adding an example on the website for the json querying, and the
json value I want to only evaluate once and then keep it around. It
would be wasteful to re-evaluate it on every keystroke. So expose a
function to return an opaque value that we can later pass back in.