Use new truncate length and format width
I just added these in the wasm module. It's quite amazing how fast it is
to render the DOM tree on every key stroke and just render spans for the
entire document, but the Vultr example is a bit long. And also, if you
write "std.range(0, 100000)" it will go haywire, stop that.

Also print slightly more narrow because the viewport tends to be quite
narrow, also due to the multi-column layout.