Add union type expressions
Using, or abusing, for now the type application syntax, the same way
that Python did it. But Python did eventually add | for union types.
I'm ambivalent about it. | is definitely more ergonomic. But for
formatting, lists with separators are always difficult to format, you
get e.g. for string literal types

    type PortVisibility =
      "public"
      | "localhost"
      | "tailnet"
      | "vpn";

which I find awkward because of the trailing ; and also the first case
is different. If we spell out union, at least those problems go away:

    type PortVisibility = Union[
      "public",
      "localhost",
      "tailnet",
      "vpn",
    ];

After writing this, I am pretty convinced that I want type application
syntax for this and not pipes. Another advantage of the more verbose
version is that it is self-documenting, you don't need to understand the
meaning of the sigils. But it is more verbose ...

I'll implement type application syntax for now, it is always possible to
change it later!