Add formatting of Union lhs type errors
I'm not really happy about the way we print this now, probably just
using the normal Atom errors is clearer anyway ... but at the top level,
it could be useful to split them out. And it's in a sense easier,
because I don't have the sourced type of the union available there.