Rewrite highlight command using MarkupString
This will enable it to be used from the webassembly module, and also
gets rid of the duplication for the ansi escape codes. I tried hard to
get the Tree-sitter highlight crate to work with webassembly, but it's
just too much of a mess and debugging linker errors is one of those
things that always takes an entire night, where you feel on the verge
of a breakthrough, but in the end it fails and you end up feeling
miserable. So no Tree-sitter wasm, maybe I can just use this
highlighter, it is less robust and less fancy, but it's better than
nothing.

This temporarily removes highlighting built-ins, maybe I can restore it
again later. Or maybe I can just highlight any function call by looking
ahead for a paren.