Clarify the DNS example
I tested it last Saturday with people at the Coffee Code & Chat meetup.
One thing that stands out is that if you don't know what Terraform is,
the example is hard to follow.

Also, that | is union, it's not obvious though people figured it out,
but we don't need it here. I might as well write the entire thing inline
and sidestep the problem.