Begin working on the smith fuzzer
For a long time it pains me that the current fuzzer works on source
code, it's wasteful, because it's unlikely that it generates e.g.
balanced brackets; it's unlikely to generate the names of builtins
without the dicteionary; and then mutation does stuff like putting the
entire source file in a string while the contents of strings in most
cases doesn't matter.

So I want to make a fuzzer where the fuzz input is some kind of
instruction language that has very little redundancies; almost every
input is valid. And then use those instructions to build up a file.
Let's see how it goes!