Handle lets in the smith fuzzer
Also typed lets. This should be enough to make it explore unions. Right
now my traditional "main" fuzzer is still searching for two code paths
in the new union types, and after hours it hasn't found it yet. I think
the new smith fuzzer should support enough by now that it could find it.
Let's find out!

Update: It found one of the code paths after just a few minutes!
Success! This is the program it synthesized:

    ╭──────╴ Opcode (hex)
    │  ╭───╴ Argument (hex)
    │  │  ╭╴ Operation, argument (decimal)
    12 ff TypePushBuiltin, 255
    12 12 TypePushBuiltin, 18
    12 08 TypePushBuiltin, 8
    34 00 ExprDictRecord, 0
    13 12 TypeApply, 18
    12 07 TypePushBuiltin, 7
    02 fe IdentPushBuiltin, 254
    20 02 ExprPushIdent, 2
    51 51 ExprTypedLet, 81
    20 02 ExprPushIdent, 2
    51 51 ExprTypedLet, 81
    -->
    let std: Union[Void,Void,] = std; let std: String = std; {}

And also the second (and final) code path that the traditional fuzzer
has not yet found, the smith fuzzer found after a few minutes:

    ╭──────╴ Opcode (hex)
    │  ╭───╴ Argument (hex)
    │  │  ╭╴ Operation, argument (decimal)
    12 01 TypePushBuiltin, 1
    02 36 IdentPushBuiltin, 54
    12 06 TypePushBuiltin, 6
    13 12 TypeApply, 18
    20 40 ExprPushIdent, 64
    20 65 ExprPushIdent, 101
    12 46 TypePushBuiltin, 70
    12 08 TypePushBuiltin, 8
    13 85 TypeApply, 133
    32 01 ExprSet, 1
    51 31 ExprTypedLet, 49
    -->
    let std: Union[Void,Set[Bool,],] = {std,}; std