Smith: Tempoarily allow only alphanumeric idents
I noticed the fuzzer started "cheating" and all the programs it
generates are just one instruction to push an ident, then one to make
it an expr, and then it spends all the effort mutating the ident input.
I want it to focus more on mutating the program. So accept only
non-interesting idents for now.